splits v2 related sdk code into it's own module
