+++ pkg/web3/sdk.go
@@ -26,15 +26,6 @@ import (
 	"github.com/lilypad-tech/lilypad/pkg/web3/bindings/storage"
 	"github.com/lilypad-tech/lilypad/pkg/web3/bindings/token"
 	"github.com/lilypad-tech/lilypad/pkg/web3/bindings/users"
-	"github.com/lilypad-tech/lilypad/pkg/web3/bindingsV2/LilypadContractRegistry"
-	"github.com/lilypad-tech/lilypad/pkg/web3/bindingsV2/LilypadModuleDirectory"
-	"github.com/lilypad-tech/lilypad/pkg/web3/bindingsV2/LilypadPaymentEngine"
-	"github.com/lilypad-tech/lilypad/pkg/web3/bindingsV2/LilypadProxy"
-	"github.com/lilypad-tech/lilypad/pkg/web3/bindingsV2/LilypadStorage"
-	"github.com/lilypad-tech/lilypad/pkg/web3/bindingsV2/LilypadToken"
-	"github.com/lilypad-tech/lilypad/pkg/web3/bindingsV2/LilypadTokenomics"
-	"github.com/lilypad-tech/lilypad/pkg/web3/bindingsV2/LilypadUser"
-	"github.com/rs/zerolog/log"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/trace"
@@ -43,7 +34,7 @@ import (
 // LP token ABI
 const erc20ABI = `[{"constant":true,"inputs":[{"name":"_owner","type":"address"}],"name":"balanceOf","outputs":[{"name":"","type":"uint256","internalType": "uint256"}],"type":"function"}]`
 
-var logger = system.GetLogger(system.Web3Service)
+var sdkLogger = system.GetLogger(system.Web3Service)
 
 // these are the go-binding wrappers for the various deployed contracts
 type Contracts struct {
@@ -57,18 +48,6 @@ type Contracts struct {
 	Pow        *pow.Pow
 }
 
-// These are the Go binding wrappers for the new version of the protocol
-type ContractsV2 struct {
-	LilypadPaymentEngine    *lilypadpaymentengine.LilypadPaymentEngine
-	LilypadProxy            *lilypadproxy.LilypadProxy
-	LilypadModuleDirectory  *lilypadmoduledirectory.LilypadModuleDirectory
-	LilypadContractRegistry *lilypadcontractregistry.LilypadContractRegistry
-	LilypadStorage          *lilypadstorage.LilypadStorage
-	LilypadUser             *lilypaduser.LilypadUser
-	LilypadTokenomics       *lilypadtokenomics.LilypadTokenomics
-	LilypadToken            *lilypadtoken.LilypadToken
-}
-
 type Web3SDK struct {
 	Options      Web3Options
 	PrivateKey   *ecdsa.PrivateKey
@@ -76,16 +55,6 @@ type Web3SDK struct {
 	CallOpts     *bind.CallOpts
 	TransactOpts *bind.TransactOpts
 	Contracts    *Contracts
-	Log          *zerolog.Logger
-}
-
-type Web3SDKV2 struct {
-	Options      Web3Options
-	PrivateKey   *ecdsa.PrivateKey
-	Client       *ethclient.Client
-	CallOpts     *bind.CallOpts
-	TransactOpts *bind.TransactOpts
-	Contracts    *ContractsV2
 	Log          *zerolog.Logger
 }
 
@@ -100,14 +69,14 @@ func NewContracts(
 	}
 
 	paymentsAddress := options.PaymentsAddress
-	logger.Debug().Msgf("paymentsAddress: %s", paymentsAddress)
+	sdkLogger.Debug().Msgf("paymentsAddress: %s", paymentsAddress)
 	if paymentsAddress == "" {
 		loadedPaymentsAddress, err := controller.GetPaymentsAddress(callOpts)
 		if err != nil {
 			return nil, err
 		}
 		paymentsAddress = loadedPaymentsAddress.String()
-		logger.Debug().
+		sdkLogger.Debug().
 			Str("load payments address", paymentsAddress).
 			Msgf("")
 	}
@@ -117,14 +86,14 @@ func NewContracts(
 	}
 
 	powAddress := options.PowAddress
-	logger.Debug().Msgf("PowAddress: %s", powAddress)
+	sdkLogger.Debug().Msgf("PowAddress: %s", powAddress)
 	if powAddress == "" {
 		loadedPowAddress, err := controller.GetPowAddress(callOpts)
 		if err != nil {
 			return nil, err
 		}
 		powAddress = loadedPowAddress.String()
-		logger.Debug().
+		sdkLogger.Debug().
 			Str("load pow address", powAddress).
 			Msgf("")
 	}
@@ -135,14 +104,14 @@ func NewContracts(
 	}
 
 	tokenAddress := options.TokenAddress
-	logger.Debug().Msgf("TokenAddress: %s", tokenAddress)
+	sdkLogger.Debug().Msgf("TokenAddress: %s", tokenAddress)
 	if tokenAddress == "" {
 		loadedTokenAddress, err := payments.GetTokenAddress(callOpts)
 		if err != nil {
 			return nil, err
 		}
 		tokenAddress = loadedTokenAddress.String()
-		logger.Debug().
+		sdkLogger.Debug().
 			Str("load token address", tokenAddress).
 			Msgf("")
 	}
@@ -153,14 +122,14 @@ func NewContracts(
 	}
 
 	storageAddress := options.StorageAddress
-	logger.Debug().Msgf("StorageAddress: %s", storageAddress)
+	sdkLogger.Debug().Msgf("StorageAddress: %s", storageAddress)
 	if storageAddress == "" {
 		loadedStorageAddress, err := controller.GetStorageAddress(callOpts)
 		if err != nil {
 			return nil, err
 		}
 		storageAddress = loadedStorageAddress.String()
-		logger.Debug().
+		sdkLogger.Debug().
 			Str("load storage address", storageAddress).
 			Msgf("")
 	}
@@ -171,14 +140,14 @@ func NewContracts(
 	}
 
 	usersAddress := options.UsersAddress
-	logger.Debug().Msgf("UsersAddress: %s", usersAddress)
+	sdkLogger.Debug().Msgf("UsersAddress: %s", usersAddress)
 	if usersAddress == "" {
 		loadedUsersAddress, err := controller.GetUsersAddress(callOpts)
 		if err != nil {
 			return nil, err
 		}
 		usersAddress = loadedUsersAddress.String()
-		logger.Debug().
+		sdkLogger.Debug().
 			Str("load users address", usersAddress).
 			Msgf("")
 	}
@@ -189,14 +158,14 @@ func NewContracts(
 	}
 
 	jobcreatorAddress := options.JobCreatorAddress
-	logger.Debug().Msgf("JobCreatorAddress: %s", jobcreatorAddress)
+	sdkLogger.Debug().Msgf("JobCreatorAddress: %s", jobcreatorAddress)
 	if jobcreatorAddress == "" {
 		loadedJobCreatorAddress, err := controller.GetJobCreatorAddress(callOpts)
 		if err != nil {
 			return nil, err
 		}
 		jobcreatorAddress = loadedJobCreatorAddress.String()
-		log.Debug().
+		sdkLogger.Debug().
 			Str("load jobcreator address", jobcreatorAddress).
 			Msgf("")
 	}
@@ -207,14 +176,14 @@ func NewContracts(
 	}
 
 	mediationAddress := options.MediationAddress
-	logger.Debug().Msgf("MediationAddress: %s", mediationAddress)
+	sdkLogger.Debug().Msgf("MediationAddress: %s", mediationAddress)
 	if mediationAddress == "" {
 		loadedMediationAddress, err := controller.GetMediationAddress(callOpts)
 		if err != nil {
 			return nil, err
 		}
 		mediationAddress = loadedMediationAddress.String()
-		log.Debug().
+		sdkLogger.Debug().
 			Str("load mediation address", mediationAddress).
 			Msgf("")
 	}
@@ -236,68 +205,10 @@ func NewContracts(
 	}, nil
 }
 
-// Creates the new Go-binding wrapper modules for the new protocol
-func NewContractsV2(
-	options Web3Options,
-	client *ethclient.Client,
-	callOpts *bind.CallOpts,
-) (*ContractsV2, error) {
-	lilypadProxy, err := lilypadproxy.NewLilypadProxy(common.HexToAddress(options.LilypadProxyAddress), client)
-	if err != nil {
-		return nil, err
-	}
-
-	lilypadModuleDirectory, err := lilypadmoduledirectory.NewLilypadModuleDirectory(common.HexToAddress(options.LilypadModuleDirectoryAddress), client)
-	if err != nil {
-		return nil, err
-	}
-
-	lilypadPaymentEngine, err := lilypadpaymentengine.NewLilypadPaymentEngine(common.HexToAddress(options.LilypadPaymentEngineAddress), client)
-	if err != nil {
-		return nil, err
-	}
-
-	lilypadContractRegistry, err := lilypadcontractregistry.NewLilypadContractRegistry(common.HexToAddress(options.LilypadContractRegistryAddress), client)
-	if err != nil {
-		return nil, err
-	}
-
-	lilypadStorage, err := lilypadstorage.NewLilypadStorage(common.HexToAddress(options.LilypadStorageAddress), client)
-	if err != nil {
-		return nil, err
-	}
-
-	lilypadUser, err := lilypaduser.NewLilypadUser(common.HexToAddress(options.LilypadUserAddress), client)
-	if err != nil {
-		return nil, err
-	}
-
-	lilypadTokenomics, err := lilypadtokenomics.NewLilypadTokenomics(common.HexToAddress(options.LilypadTokenomicsAddress), client)
-	if err != nil {
-		return nil, err
-	}
-
-	lilypadToken, err := lilypadtoken.NewLilypadToken(common.HexToAddress(options.LilypadL2TokenAddress), client)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ContractsV2{
-		LilypadPaymentEngine:    lilypadPaymentEngine,
-		LilypadProxy:            lilypadProxy,
-		LilypadContractRegistry: lilypadContractRegistry,
-		LilypadModuleDirectory:  lilypadModuleDirectory,
-		LilypadStorage:          lilypadStorage,
-		LilypadToken:            lilypadToken,
-		LilypadUser:             lilypadUser,
-		LilypadTokenomics:       lilypadTokenomics,
-	}, nil
-}
-
 func NewContractSDK(ctx context.Context, options Web3Options, tracer trace.Tracer) (*Web3SDK, error) {
 	displayOpts := options
 	displayOpts.PrivateKey = "*********"
-	logger.Debug().Msgf("NewContractSDK: %+v", displayOpts)
+	sdkLogger.Debug().Msgf("NewContractSDK: %+v", displayOpts)
 
 	client, err := getEthClient(ctx, options, tracer)
 	if err != nil {
@@ -334,51 +245,7 @@ func NewContractSDK(ctx context.Context, options Web3Options, tracer trace.Trace
 		Contracts:    contracts,
 		Log:          logger,
 	}
-	logger.Info().Msgf("Public Address: %s", web3SDK.GetAddress())
-
-	return web3SDK, nil
-}
-
-func NewContractSDKV2(ctx context.Context, options Web3Options, tracer trace.Tracer) (*Web3SDKV2, error) {
-	displayOpts := options
-	displayOpts.PrivateKey = "*********"
-	logger.Debug().Msgf("NewContractSDK: %+v", displayOpts)
-
-	client, err := getEthClient(ctx, options, tracer)
-	if err != nil {
-		return nil, err
-	}
-
-	privateKey, err := ParsePrivateKey(options.PrivateKey)
-	if err != nil {
-		return nil, err
-	}
-
-	callOpts := &bind.CallOpts{
-		Pending:     false,
-		From:        common.Address{},
-		BlockNumber: nil,
-		Context:     nil,
-	}
-
-	transactOpts, err := bind.NewKeyedTransactorWithChainID(privateKey, big.NewInt(int64(options.ChainID)))
-	if err != nil {
-		return nil, err
-	}
-	contracts, err := NewContractsV2(options, client, callOpts)
-	if err != nil {
-		return nil, err
-	}
-
-	web3SDK := &Web3SDKV2{
-		PrivateKey:   privateKey,
-		Options:      options,
-		Client:       client,
-		CallOpts:     callOpts,
-		TransactOpts: transactOpts,
-		Contracts:    contracts,
-		Log:          logger,
-	}
+	sdkLogger.Info().Msgf("Public Address: %s", web3SDK.GetAddress())
 
 	return web3SDK, nil
 }
@@ -394,7 +261,7 @@ func getEthClient(ctx context.Context, options Web3Options, tracer trace.Tracer)
 	for _, u := range rpcs {
 		parsedURL, err = url.Parse(u)
 		if err != nil {
-			logger.Warn().Msgf("Unable to parse web3 RPC URL: %v", err)
+			sdkLogger.Warn().Msgf("Unable to parse web3 RPC URL: %v", err)
 			span.RecordError(errors.New("Unable to parse web3 RPC URL"))
 			continue
 		}
@@ -402,11 +269,11 @@ func getEthClient(ctx context.Context, options Web3Options, tracer trace.Tracer)
 		span.AddEvent("ethclient.dial", trace.WithAttributes(attribute.String("web3.rpc_url", parsedURL.Host)))
 		client, err = ethclient.Dial(u)
 		if err != nil {
-			logger.Warn().Msgf("Failed to connect to %s: %v", parsedURL.Host, err)
+			sdkLogger.Warn().Msgf("Failed to connect to %s: %v", parsedURL.Host, err)
 			span.RecordError(fmt.Errorf("Failed to connect to %s", parsedURL.Host))
 			continue
 		} else {
-			logger.Info().Msgf("Connected to %s", parsedURL.Host)
+			sdkLogger.Info().Msgf("Connected to %s", parsedURL.Host)
 			span.AddEvent("ethclient.connected")
 			break
 		}
@@ -445,7 +312,7 @@ func (sdk *Web3SDK) GetBalance(address string) (*big.Int, error) {
 	// Get the balance using the converted address
 	balance, err := sdk.Client.BalanceAt(context.Background(), ethAddress, nil)
 	if err != nil {
-		logger.Error().Msgf("error for GetBalance: %s", err.Error())
+		sdkLogger.Error().Msgf("error for GetBalance: %s", err.Error())
 		return nil, err
 	}
 	return balance, nil
@@ -460,7 +327,7 @@ func (sdk *Web3SDK) GetLPBalance(address string) (*big.Int, error) {
 	// Get the balance using the converted address
 	erc20ABIObj, err := abi.JSON(strings.NewReader(erc20ABI))
 	if err != nil {
-		logger.Error().Msgf("error parsing ABI: %s", err)
+		sdkLogger.Error().Msgf("error parsing ABI: %s", err)
 		return nil, err
 	}
 	tokenContract := bind.NewBoundContract(lpToken, erc20ABIObj, client, client, client)
@@ -468,14 +335,14 @@ func (sdk *Web3SDK) GetLPBalance(address string) (*big.Int, error) {
 	var out []interface{}
 	err = tokenContract.Call(nil, &out, "balanceOf", ethAddress)
 	if err != nil {
-		logger.Error().Msgf("error calling balanceOf: %s", err)
+		sdkLogger.Error().Msgf("error calling balanceOf: %s", err)
 		return nil, err
 	}
 
 	lpBalance := *abi.ConvertType(out[0], new(big.Int)).(*big.Int)
 
 	if err != nil {
-		logger.Error().Msgf("error for GetBalance: %s", err.Error())
+		sdkLogger.Error().Msgf("error for GetBalance: %s", err.Error())
 		return nil, err
 	}
 	return &lpBalance, nil
